fix new animation calling function

+++ src/controller/extern.cpp
@@ -114,7 +114,7 @@ void Pixels_animation(Pixels* p, AnimationArgs* args) {
     p->animation(*args);
 }
 
-int* Pixels_getCurrentState(Pixels* p) {
+long* Pixels_getCurrentState(Pixels* p) {
     return p->getCurrentState();
 }
